The most significant changes involve the introduction of new methods and a new class to improve power management and user activity tracking. The `FormLoad` method in `Form1.cs` now calls a new asynchronous method `FireAndForgetAsync` instead of a debug message. This method is designed to handle exceptions and call the `MonitorAndAjustPowerPlansAsync` method from the `_powerManagementService` object.
In `PowerManagementService.cs`, the `GetCurrentPlan` method has been replaced with `MonitorAndAjustPowerPlansAsync`, which continuously gets the current power plan and writes it to the debug console every 10 seconds. A new method `SetPowerPlan` has also been added to this file, which sets the active power plan based on a provided GUID.

Additionally, a new class `UserActivityService` has been added to `UserActivityService.cs`. This class uses P/Invoke to call the `GetLastInputInfo` function from `user32.dll` to retrieve the time of the last input event.

List of changes:

1. `FormLoad` method in `Form1.cs` has been modified to call a new asynchronous method `FireAndForgetAsync` instead of a debug message.
2. `GetCurrentPlan` method in `PowerManagementService.cs` has been replaced with `MonitorAndAjustPowerPlansAsync`.
3. A new method `SetPowerPlan` has been added to `PowerManagementService.cs`.
4. `GetCurrentPlan` method in `PowerManagementService.cs` has been moved within the file.
5. A new class `UserActivityService` has been added to `UserActivityService.cs` to retrieve the time of the last input event.